% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celfunc-package.R
\docType{package}
\name{celfunc-package}
\alias{celfunc}
\alias{celfunc-package}
\title{celfunc: Celon Pharma Utility Functions}
\description{
A collection of small utility functions for PK/PD data wrangling,
formatting, and list traversal used in CPL'36 analyses at Celon Pharma S.A.
}
\author{
\strong{Maintainer}: Radek Jaźwiec \email{radoslaw.jazwiec@celonpharma.com} (\href{https://orcid.org/0000-0002-9272-6233}{ORCID})

}
\keyword{internal}
