% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mean.R
\name{geom_mean}
\alias{geom_mean}
\title{Geometric mean of a numeric vector}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{Numeric vector. \code{NA} values are removed before computation.}
}
\value{
A single numeric value, or \code{NA} if any non-missing value is \verb{<= 0}.
}
\description{
Geometric mean of a numeric vector
}
\examples{
geom_mean(c(1, 4, 16))   # 4
geom_mean(c(1, -1, 4))   # NA

}
